requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.3
chromadb
sentence-transformers
duckduckgo-search
gTTS
python-dotenv
pypdf
Pillow
watchdog
numpy<2.0
tiktoken